+++ app/helpers/planning_elements_helper.rb
@@ -43,8 +43,8 @@ module PlanningElementsHelper
         api.responsible(:id => planning_element.responsible.id, :name => planning_element.responsible.name)
       end
 
-      if planning_element.planning_element_type
-        type = planning_element.planning_element_type
+      if planning_element.type
+        type = planning_element.type
         api.planning_element_type(:id => type.id, :name => type.name)
       end