+++ features/step_definitions/general_steps.rb
@@ -263,12 +263,15 @@ Given /^the [pP]roject "([^\"]*)" has 1 [sS]ubproject with the following:$/ do |
 end
 
 Given /^there are the following types:$/ do |table|
-
+  table.map_headers! { |header| header.underscore.gsub(' ', '_') }
   table.hashes.each_with_index do |t, i|
     type = Type.find_by_name(t['name'])
     type = Type.new :name => t['name'] if type.nil?
-    type.position = t['position'] ? t['position'] : i
-    type.is_in_roadmap = t['is_in_roadmap'] ? t['is_in_roadmap'] : true
+    type.position       = t['position'] ? t['position'] : i
+    type.is_in_roadmap  = t['is_in_roadmap'] ? t['is_in_roadmap'] : true
+    type.is_milestone   = t['is_milestone'] ? t['is_milestone'] : true
+    type.is_default     = t['is_default'] ? t['is_default'] : false
+    type.in_aggregation = t['in_aggregation'] ? t['in_aggregation'] : true
     type.save!
   end
 end